+++ sprite.cpp
@@ -11,7 +11,7 @@ Sprite::Sprite(Graphics& graph, std::string spritePath,int width,int height)
     this->sourceRect = {0,0,width,height};
 }
 
-void Sprite::draw(Graphics& graph,int x,int y)
+void Sprite::draw(Graphics& graph,float x,float y)
 {
     SDL_Rect dest_rect = { x,y, this->sourceRect.w, this->sourceRect.h};
     graph.blitSurface(this->texture,&dest_rect);
